+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/lowlevel/query/impl/FieldContextSimpleQueryParser.java
@@ -5,7 +5,7 @@
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
 
-package org.hibernate.search.backend.lucene.util.impl;
+package org.hibernate.search.backend.lucene.lowlevel.query.impl;
 
 import java.util.Collections;
 import java.util.Map;
