+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/search/extraction/impl/HibernateSearchDocumentIdToLuceneDocIdMapCollector.java
@@ -10,7 +10,7 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
-import org.hibernate.search.backend.lucene.util.impl.LuceneFields;
+import org.hibernate.search.backend.lucene.lowlevel.common.impl.LuceneFields;
 
 import org.apache.lucene.index.BinaryDocValues;
 import org.apache.lucene.index.DocValues;
