+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/types/codec/impl/LuceneGeoPointFieldCodec.java
@@ -6,12 +6,12 @@
  */
 package org.hibernate.search.backend.lucene.types.codec.impl;
 
-import static org.hibernate.search.backend.lucene.util.impl.LuceneFields.internalFieldName;
+import static org.hibernate.search.backend.lucene.lowlevel.common.impl.LuceneFields.internalFieldName;
 
 import java.util.function.Consumer;
 
 import org.hibernate.search.backend.lucene.document.impl.LuceneDocumentBuilder;
-import org.hibernate.search.backend.lucene.util.impl.LuceneFields;
+import org.hibernate.search.backend.lucene.lowlevel.common.impl.LuceneFields;
 import org.hibernate.search.engine.spatial.GeoPoint;
 
 import org.apache.lucene.document.Document;
