+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/types/sort/comparatorsource/impl/LuceneGeoPointDistanceComparatorSource.java
@@ -8,9 +8,9 @@ package org.hibernate.search.backend.lucene.types.sort.comparatorsource.impl;
 
 import java.io.IOException;
 
-import org.hibernate.search.backend.lucene.types.lowlevel.impl.GeoPointDistanceDocValues;
-import org.hibernate.search.backend.lucene.types.lowlevel.impl.OnTheFlyNestedSorter;
-import org.hibernate.search.backend.lucene.types.lowlevel.impl.SortedNumericDoubleValues;
+import org.hibernate.search.backend.lucene.lowlevel.docvalues.impl.GeoPointDistanceDocValues;
+import org.hibernate.search.backend.lucene.lowlevel.sort.impl.OnTheFlyNestedSorter;
+import org.hibernate.search.backend.lucene.lowlevel.docvalues.impl.SortedNumericDoubleValues;
 
 import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.index.NumericDocValues;
