+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/types/sort/comparatorsource/impl/LuceneTextFieldComparatorSource.java
@@ -8,7 +8,8 @@ package org.hibernate.search.backend.lucene.types.sort.comparatorsource.impl;
 
 import java.io.IOException;
 
-import org.hibernate.search.backend.lucene.types.lowlevel.impl.OnTheFlyNestedSorter;
+import org.hibernate.search.backend.lucene.lowlevel.sort.impl.OnTheFlyNestedSorter;
+import org.hibernate.search.backend.lucene.lowlevel.docvalues.impl.LuceneReplaceMissingSortedDocValues;
 import org.hibernate.search.backend.lucene.types.sort.impl.SortMissingValue;
 
 import org.apache.lucene.index.LeafReaderContext;
