+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/work/impl/LuceneQueryBasedUpdateEntryWork.java
@@ -10,7 +10,7 @@ import java.io.IOException;
 
 import org.hibernate.search.backend.lucene.document.impl.LuceneIndexEntry;
 import org.hibernate.search.backend.lucene.lowlevel.writer.impl.IndexWriterDelegator;
-import org.hibernate.search.backend.lucene.search.impl.LuceneQueries;
+import org.hibernate.search.backend.lucene.lowlevel.query.impl.LuceneQueries;
 
 
 public class LuceneQueryBasedUpdateEntryWork extends AbstractLuceneUpdateEntryWork {
